fix: incorrectly skipping downloading

+++ src/DownloadWebGallery.py
@@ -3,6 +3,7 @@ import os
 import shutil
 import sqlite3
 import sys
+import hashlib
 
 from bs4 import BeautifulSoup
 from loguru import logger
@@ -33,23 +34,32 @@ class DownloadWebGallery(Config):
         self.filepath_end = os.path.join(self.data_path, folder, str(self.gid) + '-' + self.title + "-1280x")
 
         self.long_url = f"https://{self.base_url}/g/{self.gid}/{self.token}/"
-
+    
+    async def calc_sha256(self, file_path):
+        sha256 = hashlib.sha256()
+        with open(file_path, 'rb') as f:
+            for chunk in iter(lambda: f.read(4096), b''):
+                sha256.update(chunk)
+        return sha256.hexdigest()
+    
     @logger.catch()
     async def download_image(self, semaphore, url, file_path):
         async with semaphore:
             real_url = await self.fetch_data(url=url)
             real_url = BeautifulSoup(real_url, 'html.parser')
             real_url = real_url.select_one('img#img').get('src')
+            # 配额用尽，返回的real_url会变成509.gif
             if real_url == "https://exhentai.org/img/509.gif":
                 logger.warning("509: YOU HAVE TEMPORARILY REACHED THE LIMIT")
                 if self.get_image_limits() == False:
-                        logger.warning("Can't get image limits as the account has been suspended, Attempt to wait for 12 hours")
-                        # 账号已被暂停，无法访问配额页面，等待12小时配额自行恢复 / The account has been suspended
-                        time.sleep(12 * 60 * 60)
+                    # 账号已被暂停，无法访问配额页面，等待12小时配额自行恢复 / The account has been suspended
+                    logger.warning("Can't get image limits as the account has been suspended, Attempt to wait for 12 hours")
+                    time.sleep(12 * 60 * 60)
                 else:
                     image_limits,total_limits=self.get_image_limits()
                     logger.warning(f"Currently at {image_limits} towards the limit of {total_limits}. Waiting for quota restoration")
                     while True:
+                        # 每隔一小时检查配额是否恢复 / Check every hour to see if the quota is restored
                         time.sleep(3600)
                         image_limits,_=self.get_image_limits()
                         print(f"Currently at {image_limits}")
@@ -57,8 +67,23 @@ class DownloadWebGallery(Config):
                             break
                 return await self.download_image(semaphore, url, file_path)
             file = await self.fetch_data(url=real_url)
-            with open(file_path, 'wb') as f:
-                f.write(file)
+            if os.path.exists(file_path):
+                # 如果已存在该文件则进行哈希校验判断是否需要下载
+                # If the file already exists, perform a hash check to determine whether it needs to be downloaded
+                exist_hash = await self.calc_sha256(file_path)
+                base_name, ext = os.path.splitext(file_path)
+                counter = 2
+                while os.path.exists(f"{base_name} ({counter}){ext}"):
+                    counter += 1
+                file_path = f"{base_name} ({counter}){ext}"
+                with open(file_path, 'wb') as f:
+                    f.write(file)
+                new_hash = self.calc_sha256(file_path)
+                if new_hash == exist_hash:
+                    os.remove(file_path)
+            else:
+                with open(file_path, 'wb') as f:
+                    f.write(file)
 
     @logger.catch()
     async def get_image_url(self):
@@ -82,6 +107,14 @@ class DownloadWebGallery(Config):
         if copyright_msg is not None:
             if copyright_msg.find('copyright') != -1:
                 return "copyright"
+        
+        # 获取页面上显示的pages，用于校验是否获取到全部的page_img_url
+        pages = None
+        lengthtd = page_data.find('td', text='Length:')
+        if lengthtd:
+            length = lengthtd.find_next_sibling('td', class_='gdt2').text.strip()
+            pages = re.search(r'\d+', length).group()
+            pages=int(pages)
 
         # 获取总页码 / Get the total page number
         ptb = page_data.select_one('.ptb td:nth-last-child(2)')
@@ -114,18 +147,24 @@ class DownloadWebGallery(Config):
             sub_ptb = sub_ptb + 1
             
         if len(page_img_url)==0:
-            # 加载到https://exhentai.org/img/blank.gif
+            # 加载到https://exhentai.org/img/blank.gif，导致获取链接为空
             logger.warning("Failed to get image urls, retrying after 30mins……")
             time.sleep(30 * 60)
             return await self.get_image_url()
+        elif pages and len(page_img_url) < pages:
+            # 检查是否获取到全部链接
+            logger.warning("Failed for missing pages, retrying after 30mins……")
+            time.sleep(30 * 60)
+            return await self.get_image_url()
 
-        return page_img_url
+        # 返回pages用于移动时重复校验
+        return page_img_url, pages
 
     @logger.catch()
     async def apply(self):
         logger.info(f"Download Web Gallery...: {self.long_url}")
 
-        img_url_res = await self.get_image_url()
+        img_url_res, pages = await self.get_image_url()
 
         if img_url_res == "copyright":
             logger.warning(
@@ -142,16 +181,19 @@ class DownloadWebGallery(Config):
         for _p in img_url_res:
             url = _p[0]
             file_path = os.path.join(self.filepath_tmp, _p[1])
-            if os.path.exists(file_path):
-                logger.info(f"File {_p[1]} already exists. Skipping download.")
-            else:
-                req = self.download_image(semaphore, url, file_path)
-                task_list.append(req)
+            # if os.path.exists(file_path):
+            #     logger.info(f"File {_p[1]} already exists. Skipping download.")
+            # else:
+            #     req = self.download_image(semaphore, url, file_path)
+            #     task_list.append(req)
+            # 部分画廊的图片文件名存在相同的情况，跳过下载会导致缺页
+            req = self.download_image(semaphore, url, file_path)
+            task_list.append(req)
 
         await tqdm_asyncio.gather(*task_list)
 
         # 判断 new_path 是否存在文件夹
-        # # Determine whether there is a folder in new_path
+        # Determine whether there is a folder in new_path
         check_dir = None
         if os.path.isdir(self.filepath_end):
             logger.warning("Directory already exists, Whether to delete the overlay? Y/N\n")
@@ -169,6 +211,19 @@ class DownloadWebGallery(Config):
                     check_dir = None
         else:
             shutil.move(self.filepath_tmp, self.filepath_end)
+            
+        # 再次检查是否缺页 / Recheck for missing pages
+        file_count = 0
+        for _, _, files in os.walk(self.filepath_end):
+            file_count += len(files)
+        if pages:
+            if file_count < pages:
+                logger.warning(f"Failed for missing pages: {self.long_url}")
+                return False
+        else:
+            if file_count < len(img_url_res):
+                logger.warning(f"Failed for missing pages: {self.long_url}")
+                return False
 
         with sqlite3.connect(self.dbs_name) as co:
             co.execute(f'UPDATE fav_category SET web_1280x_flag = 1 WHERE gid = {self.gid}')
